observability/opentelemetry-elastic/elastiflix/python-favorite-otel-manual/requirements.txt
async-timeout==4.0.2
blinker==1.6.2
click==8.1.3
ecs-logging==2.0.2
Flask==2.3.2
importlib-metadata~=6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
redis==4.5.5
Werkzeug==2.3.6
zipp==3.15.0
opentelemetry-api==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-flask==0.39b0
opentelemetry-instrumentation-requests==0.39b0
opentelemetry-instrumentation-wsgi==0.39b0
opentelemetry-instrumentation-redis==0.39b0
opentelemetry-exporter-otlp==1.18.0
